requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.1
openai==1.12.0
SpeechRecognition==3.10.1
pyttsx3==2.90
python-engineio==4.9.0
python-socketio==5.11.1 